Renamed order folder and Order file to orders-page and DashBoardOrdersPage. Renamed product-catalog to product-catalog-page and ProductCatalog to ProductCatalogPage
