+++ ecosphere-goods/src/components/pages/dashboard/product-catalog-page/ProductCatalogDisplay.js
