+++ ecosphere-goods/src/components/pages/dashboard/product-catalog-page/ProductCatalogItem.js
