+++ ecosphere-goods/src/components/pages/dashboard/product-catalog-page/ProductCatalogPage.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import ProductCatalogDisplay from './ProductCatalogDisplay'
 import ProductDialog from './product-handling/ProductDialog'
-import { ProductCatalogProvider } from '../product-catalog/ProductCatalogContext'
+import { ProductCatalogProvider } from './ProductCatalogContext'
 import PaginationProvider from '../../../utility/pagination/PaginationContext'
 
-const ProductCatalog = () => {
+const ProductCatalogPage = () => {
 
   return (
     <ProductCatalogProvider>
@@ -24,4 +24,4 @@ const ProductCatalog = () => {
   )
 }
 
-export default ProductCatalog
+export default ProductCatalogPage
