+++ ecosphere-goods/src/components/pages/dashboard/product-catalog-page/product-handling/ConfirmDeleteDialog.js
